Removed references to winget in install-prerequisistes
WinGet is not installed on GitHub runners (CI workflows). The method
to force WinGet installation on GitHub runners failed from time to
time, leading to workflow failure. Recently, the number of failures
was so high that it is more reasonable to avoid using WinGet in
GitHub runners. All installation scripts which used WinGet were
modified to fetch the installers from their source.